refactor: ProductManagement 컴포넌트 분리
